% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past_starmap.R
\name{plot_starmap}
\alias{plot_starmap}
\title{Plot A Custom Starmap}
\usage{
plot_starmap(
  location,
  date = today(),
  style = "black",
  line1_text = "",
  line2_text = "",
  line3_text = ""
)
}
\arguments{
\item{location}{Location of choice. Uses the ArGis geocoding API through via `tidygeocoder::geocode`. Used to extract latitude}

\item{date}{date. ymd format works best, other formats have not been tested.}

\item{style}{choice. Right now only accepts "black and green}

\item{line1_text}{Text you want to have in your image caption. There are 3 centered lines specified.}

\item{line2_text}{Ibid.}

\item{line3_text}{Ibid.}
}
\description{
Create a custom, print quality star map with `plot_starmap()`
}
\examples{

plot_starmap(location= "Toronto, ON, Canada",
             date="2022-01-17",
             style="black",
             line1_text="Toronto",
             line2_text ="January 17th, 2023",
             line3_text="43.6532째 N, 79.3832째 W")

plot_starmap(location= "Toronto, ON, Canada",
             date="2022-01-17",
             style="green",
             line1_text="Toronto",
             line2_text ="January 17th, 2023",
             line3_text="43.6532째 N, 79.3832째 W")
}
